LRAC-5823 analytics-message-sender-api Move EntityModelListener to appropriate package
