+++ modules/apps/analytics/analytics-message-sender-api/src/main/java/com/liferay/analytics/message/sender/model/listener/BaseEntityModelListener.java
@@ -15,7 +15,6 @@
 package com.liferay.analytics.message.sender.model.listener;
 
 import com.liferay.analytics.message.sender.model.AnalyticsMessage;
-import com.liferay.analytics.message.sender.model.EntityModelListener;
 import com.liferay.analytics.message.sender.util.AnalyticsExpandoBridgeUtil;
 import com.liferay.analytics.message.storage.service.AnalyticsMessageLocalService;
 import com.liferay.analytics.settings.configuration.AnalyticsConfiguration;
